requirements.txt
# Python 3.11+ 
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.10.0

# Data Scraping
playwright>=1.35.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# UI
streamlit>=1.25.0
plotly>=5.15.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
loguru>=0.7.0
